Make clicking on an score td always open the input
Two clicks are needed now - one to close the input, and another to open
the next input. This is because opening and closing an input moves the
score cells right and left. Fix this by making the input wait 1/10 of a
second to close, which should be enough for the click to process.