Don't imply the null score when score.number === 0
Currently, when the user enters a score of 0, the scorecard treats that
like the null score and doesn't display anything, which can be
confusing.

When the user enters in the number 0, a 0 should show up, but when they
enter nothing, nothing should show up (i.e. the null score). When they
enter only a symbol, that symbol should show up, but no number.

In other words, make sure to display as close to what the user entered
as possible.