Move the Add Player and Save buttons below scores
This involves adding two rows to the table footer.  The first row is for the Add Player button and the Save button.  The second row is after the Notes row and is for the Scorecard ID, which I added as well.  Joel won't like that, but committing every small change when there are already other changes to a set of files is not easy nor always practical.  TODO: The Save button still needs to be expanded in width, but it doesn't obey the CSS width property.