Don't have a separate scorecard for each course
Instead, make the course a property of the scorecard. This will allow
users to save the course with the scorecard and to change the par. It
also may help with keeping the code clean.

This will involve changes throughout the scorecard.js file, since it
will change the structure of the JSON. It also will mean that the
currentCourse variable won't be necessary anymore.

Use courseId() rather than courseStr() to identify courses. Currently,
calling .val() on the course select returns the value of courseStr(),
but this change will set the value attribute on each <option>, so that
the course ID can be used for identification but the user can still see
the result of courseStr().

Also try to align variable assignments at the top of functions when it
seems to help, and functions to not rely on global variables when
possible (e.g. addPlayer() currently relies on a global, but this will
change that).